Calculate a vector of day lengths for a given latitude, by julian day

Calculate a vector of day lengths for a given latitude, by julian day

Usage

daylengths(latitude, precision = 64)

Arguments

latitude

Latitude of location

precision

Optional, precision of day integral (default value 64)

Value

Vector of day lengths in hours

Examples

daylengths(-43)
#>   [1] 15.140625 15.109375 15.109375 15.078125 15.078125 15.046875 15.046875
#>   [8] 15.015625 15.015625 14.984375 14.953125 14.921875 14.921875 14.890625
#>  [15] 14.859375 14.828125 14.796875 14.765625 14.734375 14.703125 14.671875
#>  [22] 14.640625 14.609375 14.578125 14.546875 14.515625 14.484375 14.453125
#>  [29] 14.390625 14.359375 14.328125 14.296875 14.234375 14.203125 14.171875
#>  [36] 14.109375 14.078125 14.046875 13.984375 13.953125 13.921875 13.859375
#>  [43] 13.828125 13.765625 13.734375 13.671875 13.640625 13.578125 13.546875
#>  [50] 13.484375 13.453125 13.390625 13.359375 13.296875 13.265625 13.203125
#>  [57] 13.171875 13.109375 13.078125 13.015625 12.984375 12.921875 12.859375
#>  [64] 12.828125 12.765625 12.734375 12.671875 12.640625 12.578125 12.515625
#>  [71] 12.484375 12.421875 12.390625 12.328125 12.265625 12.234375 12.171875
#>  [78] 12.140625 12.078125 12.046875 11.984375 11.921875 11.890625 11.828125
#>  [85] 11.796875 11.734375 11.671875 11.640625 11.578125 11.546875 11.484375
#>  [92] 11.453125 11.390625 11.359375 11.296875 11.234375 11.203125 11.140625
#>  [99] 11.109375 11.046875 11.015625 10.953125 10.921875 10.859375 10.828125
#> [106] 10.765625 10.734375 10.671875 10.640625 10.578125 10.546875 10.515625
#> [113] 10.453125 10.421875 10.359375 10.328125 10.265625 10.234375 10.203125
#> [120] 10.140625 10.109375 10.078125 10.015625  9.984375  9.953125  9.890625
#> [127]  9.859375  9.828125  9.796875  9.734375  9.703125  9.671875  9.640625
#> [134]  9.609375  9.546875  9.515625  9.484375  9.453125  9.421875  9.390625
#> [141]  9.359375  9.328125  9.296875  9.265625  9.234375  9.203125  9.171875
#> [148]  9.171875  9.140625  9.109375  9.078125  9.078125  9.046875  9.015625
#> [155]  9.015625  8.984375  8.953125  8.953125  8.921875  8.921875  8.890625
#> [162]  8.890625  8.890625  8.859375  8.859375  8.859375  8.828125  8.828125
#> [169]  8.828125  8.828125  8.828125  8.828125  8.828125  8.828125  8.828125
#> [176]  8.828125  8.828125  8.828125  8.828125  8.859375  8.859375  8.859375
#> [183]  8.859375  8.890625  8.890625  8.921875  8.921875  8.953125  8.953125
#> [190]  8.984375  8.984375  9.015625  9.046875  9.046875  9.078125  9.109375
#> [197]  9.109375  9.140625  9.171875  9.203125  9.234375  9.265625  9.296875
#> [204]  9.328125  9.359375  9.390625  9.421875  9.453125  9.484375  9.515625
#> [211]  9.546875  9.578125  9.609375  9.640625  9.671875  9.734375  9.765625
#> [218]  9.796875  9.828125  9.890625  9.921875  9.953125  9.984375 10.046875
#> [225] 10.078125 10.109375 10.171875 10.203125 10.234375 10.296875 10.328125
#> [232] 10.390625 10.421875 10.453125 10.515625 10.546875 10.609375 10.640625
#> [239] 10.703125 10.734375 10.796875 10.828125 10.890625 10.921875 10.984375
#> [246] 11.015625 11.078125 11.109375 11.171875 11.203125 11.265625 11.296875
#> [253] 11.359375 11.390625 11.453125 11.484375 11.546875 11.578125 11.640625
#> [260] 11.671875 11.734375 11.765625 11.828125 11.890625 11.921875 11.984375
#> [267] 12.015625 12.078125 12.109375 12.171875 12.203125 12.265625 12.328125
#> [274] 12.359375 12.421875 12.453125 12.515625 12.546875 12.609375 12.640625
#> [281] 12.703125 12.765625 12.796875 12.859375 12.890625 12.953125 12.984375
#> [288] 13.046875 13.078125 13.140625 13.171875 13.234375 13.265625 13.328125
#> [295] 13.359375 13.421875 13.453125 13.515625 13.546875 13.609375 13.640625
#> [302] 13.703125 13.734375 13.765625 13.828125 13.859375 13.921875 13.953125
#> [309] 13.984375 14.046875 14.078125 14.109375 14.171875 14.203125 14.234375
#> [316] 14.296875 14.328125 14.359375 14.390625 14.421875 14.484375 14.515625
#> [323] 14.546875 14.578125 14.609375 14.640625 14.671875 14.703125 14.734375
#> [330] 14.765625 14.796875 14.828125 14.859375 14.890625 14.890625 14.921875
#> [337] 14.953125 14.984375 14.984375 15.015625 15.046875 15.046875 15.078125
#> [344] 15.078125 15.109375 15.109375 15.140625 15.140625 15.140625 15.171875
#> [351] 15.171875 15.171875 15.171875 15.171875 15.171875 15.171875 15.171875
#> [358] 15.171875 15.171875 15.171875 15.171875 15.171875 15.171875 15.140625
#> [365] 15.140625 15.140625