Process Vegetation polygon data to classified rasters

Process Vegetation polygon data to classified rasters

Usage

processVeg(
  outdir,
  veg_polygon,
  veg_LUT,
  join_field,
  min_field = "MIN",
  max_field = "MAX",
  form_field = NULL,
  quiet = TRUE
)

Arguments

outdir

Directory to output vegetation maps, must contains files produced by createDomain

veg_polygon

SpatVect or SimpleFeatures object containing the vegetation map. Must contain a field with unique vegetation ID.

veg_LUT

Tibble or data.frame containing look-up table linking veg ID to MAX and MIN intervals

join_field

Name fo field joining veg_polygon and veg_LUT

min_field

Field name in look-up table containing minimum interval in years

max_field

Field name in look-up table containing maximum interval in years

form_field

(Optional) Field name in look-up table containing the broad vegetation formation or group to summarise results within.

quiet

Verbose logging

Value

Path to outdir for piping.

Examples

if (FALSE) { # \dontrun{

data("COH_Veg")
COH_Veg <- vect(COH_Veg)
data("COH_Veg_LUT")
processVeg("output",veg_polygon=COH_Veg,veg_LUT=COH_Veg_LUT,
    join_field="VEGCODE",form_field="VEG_GROUP")
} # }